+++ packages/resume/changelog.md
@@ -1,6 +1,8 @@
 # Unreleased
 * Add combinators `resumeOr` and `resumingOr`, which take an additional branch for the success case.
-* Add operator versions of `resume` (`!!`) and `resumeAs` (`<!`).
+* Add operator versions of `resume` (`!!`) and `resumeAs` (`<!`), (`!>`).
+* Add combinators `resumeWith` and `resumingWith` that ignore the error and execute an action, plus their operator
+  variants `(!>>)` and `(<<!)`.
 
 # 0.1.0.0
 * initial release